+++ examples/7.Display/LCD_5110/Terminal12.h
@@ -1,5 +1,5 @@
 //
-//  Terminal11.h
+//  Terminal12.h
 //  Font library
 //  ----------------------------------
 //  Developed with embedXcode
