+++ examples/7.Display/LCD_BoosterPack/Terminal12.h
@@ -1,3 +1,23 @@
+///
+/// @file	Terminal11.h
+/// @brief	Font library
+/// @details	Font Terminal 12 x 16 
+/// @n	
+/// @n	
+/// @n @a	Developed with [embedXcode](http://embedXcode.weebly.com)
+/// @n
+/// @author	Rei VILO
+/// @author	http://embeddedcomputing.weebly.com
+/// @date	May 26, 2012
+/// @version	1
+/// @n
+/// @copyright	© Rei VILO, 2012
+/// @copyright	CC = BY NC SA
+///
+/// @see	Font Generated by MikroElektronika GLCD Font Creator 1.2.0.0
+/// @n	 	MikroeElektronika 2011 http://www.mikroe.com 
+///
+
 //
 //  Terminal11.h
 //  Font library
@@ -21,6 +41,17 @@
 // GLCD FontSize : 11 x 16
 //
 
+///
+/// @brief	Font Terminal11x16
+/// @details 	GLCD FontName : Terminal11x16
+/// @n 		GLCD FontSize : 11 x 16
+///
+/// @note 	WARNING: This Font is usable only with MikroE GLCD Lib.
+/// @n		X-GLCD Lib does not handle this font.
+///
+/// @copyright	Font Generated by MikroElektronika GLCD Font Creator 1.2.0.0
+/// @see 	MikroeElektronika 2011 http://www.mikroe.com 
+///
 static const unsigned char Terminal11x16[96][22] = {
   0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00, // Code for char  
   0x00,0x00,0x00,0x00,0x00,0x00,0x7C,0x00,0xFF,0x33,0xFF,0x33,0x7C,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00, // Code for char !
