+++ examples/7.Display/LCD_BoosterPack/Terminal6.h
@@ -1,3 +1,23 @@
+///
+/// @file	Terminal6.h
+/// @brief	Font library
+/// @details	Font Terminal 6 x 8 
+/// @n	
+/// @n	
+/// @n @a	Developed with [embedXcode](http://embedXcode.weebly.com)
+/// @n
+/// @author	Rei VILO
+/// @author	http://embeddedcomputing.weebly.com
+/// @date	May 26, 2012
+/// @version	1
+/// @n
+/// @copyright	© Rei VILO, 2012
+/// @copyright	CC = BY NC SA
+///
+/// @see	Font Generated by MikroElektronika GLCD Font Creator 1.2.0.0
+/// @n	 	MikroeElektronika 2011 http://www.mikroe.com 
+///
+
 //
 //  Terminal6.h
 //  Font library
@@ -21,6 +41,17 @@
 // GLCD FontSize : 5 x 8
 //
 
+///
+/// @brief	Font Terminal6x8
+/// @details 	GLCD FontName : Terminal6x8
+/// @n 		GLCD FontSize : 5 x 8
+///
+/// @note 	WARNING: This Font is usable only with MikroE GLCD Lib.
+/// @n		X-GLCD Lib does not handle this font.
+///
+/// @copyright	Font Generated by MikroElektronika GLCD Font Creator 1.2.0.0
+/// @see 	MikroeElektronika 2011 http://www.mikroe.com 
+///
 static const unsigned char Terminal6x8[96][5] = {
   0x00, 0x00, 0x00, 0x00, 0x00,      // Code for char  
   0x00, 0x06, 0x5F, 0x06, 0x00,      // Code for char !
